+++ dzgui/views/dialogs/servers.py
@@ -44,8 +44,8 @@ class ServerDialog(GenericDialog):
         self.content = self.get_content_area()
         self.content.pack_start(self.scrollable_tree, EXPAND, FILL, 0)
 
-    def pack(self, widget: Gtk.Widget) -> None:
-        self.content.pack_start(widget, EXPAND, FILL, NO_PADDING)
+    # def pack(self, widget: Gtk.Widget) -> None:
+    #     self.content.pack_start(widget, EXPAND, FILL, NO_PADDING)
 
     def _on_keypress(self, view: Gtk.TreeView, event: Gdk.EventKey) -> None:
         # NOTE: ESC normally unfocuses treeview instead of destroying dialog